Refactored packages and viewholders separately
