+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/FurnitureMainFragmentPresenter.java
@@ -2,7 +2,8 @@ package com.example.hsh.homesweethome.furniture.presenter;
 
 import com.example.hsh.homesweethome.Models.CategoryFurniture;
 import com.example.hsh.homesweethome.Models.Furniture;
-import com.example.hsh.homesweethome.furniture.view.IFurnitureMainFragmentView;
+import com.example.hsh.homesweethome.furniture.presenter.interfaces.IFurnitureMainFragmentPresenter;
+import com.example.hsh.homesweethome.furniture.view.interfaces.IFurnitureMainFragmentView;
 import com.example.hsh.homesweethome.network.APIService;
 import com.example.hsh.homesweethome.network.Utils;
