+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/interfaces/IFragmentNavigationPresenter.java
@@ -1,0 +1,7 @@
+package com.example.hsh.homesweethome.furniture.presenter.interfaces;
+
+import com.example.hsh.homesweethome.furniture.view.FurnitureBaseFragment;
+
+public interface IFragmentNavigationPresenter {
+    void changeFragment(FurnitureBaseFragment fragment);
+}
