+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/interfaces/IRecyclerViewAdapterFurniturePresenter.java
@@ -1,0 +1,8 @@
+package com.example.hsh.homesweethome.furniture.presenter.interfaces;
+
+import com.example.hsh.homesweethome.furniture.view.interfaces.RecyclerViewHolderFurniture;
+
+public interface IRecyclerViewAdapterFurniturePresenter {
+    void onBindFurnitureViewAtPosition(int position, RecyclerViewHolderFurniture viewHolder);
+    int getFurnitureCount();
+}
