+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/interfaces/IRecyclerViewAdapterMainPresenter.java
@@ -1,6 +1,6 @@
-package com.example.hsh.homesweethome.furniture.presenter;
+package com.example.hsh.homesweethome.furniture.presenter.interfaces;
 
-import com.example.hsh.homesweethome.furniture.RecyclerViewHolderCategoryFurniture;
+import com.example.hsh.homesweethome.furniture.view.interfaces.RecyclerViewHolderCategoryFurniture;
 
 public interface IRecyclerViewAdapterMainPresenter {
     void onBindFurnitureCategoryAtPosition(int position, RecyclerViewHolderCategoryFurniture viewHolder);
