--- app/src/main/java/com/example/hsh/homesweethome/furniture/view/IFragmentNavigationView.java
@@ -1,7 +1,0 @@
-package com.example.hsh.homesweethome.furniture.view;
-
-import com.example.hsh.homesweethome.furniture.presenter.IFragmentNavigationPresenter;
-
-public interface IFragmentNavigationView {
-    void attachPresenter(IFragmentNavigationPresenter presenter);
-}
