--- app/src/main/java/com/example/hsh/homesweethome/furniture/view/IFurnitureView.java
@@ -1,7 +1,0 @@
-package com.example.hsh.homesweethome.furniture.view;
-
-import com.example.hsh.homesweethome.furniture.FurnitureBaseFragment;
-
-public interface IFurnitureView {
-    void setFragment(FurnitureBaseFragment fragment);
-}
