+++ app/src/main/java/com/example/hsh/homesweethome/furniture/view/interfaces/IFragmentNavigationView.java
@@ -1,0 +1,7 @@
+package com.example.hsh.homesweethome.furniture.view.interfaces;
+
+import com.example.hsh.homesweethome.furniture.presenter.interfaces.IFragmentNavigationPresenter;
+
+public interface IFragmentNavigationView {
+    void attachPresenter(IFragmentNavigationPresenter presenter);
+}
