+++ app/src/main/java/com/example/hsh/homesweethome/furniture/view/interfaces/IFurnitureView.java
@@ -1,0 +1,7 @@
+package com.example.hsh.homesweethome.furniture.view.interfaces;
+
+import com.example.hsh.homesweethome.furniture.view.FurnitureBaseFragment;
+
+public interface IFurnitureView {
+    void setFragment(FurnitureBaseFragment fragment);
+}
