+++ app/src/main/java/com/example/hsh/homesweethome/furniture/view/interfaces/RecyclerViewHolderLocation.java
@@ -1,0 +1,5 @@
+package com.example.hsh.homesweethome.furniture.view.interfaces;
+
+public interface RecyclerViewHolderLocation {
+    void setLocation(String location);
+}
